Fix check if NegotiatedProtocol is empty
If peer doesn't support ALPN, ConnectionState.NegotiatedProtocol will be empty.
NegotiatedProtocol should only be passed when not empty otherwise connection will
fail. We pass what the client supports to the server and look at what was negotiated
with the server. On the local end, we pick the exact protocol so that ALPN works.